ref: extract pack stuff out of ModCore
